vehicle, start and finish lines, and item base classes
Added a Vehicle base class with some variables and methods. No functionality yet.
Also added start line and finish line classes. they're empty.
Finally, added Pickup Actor and Item base class as well as an empty item for when the player has no items. The ItemBase and empty item are parent/child and the pickup item adds an item to the vehicle as  a component to allow for dynamic items.