More reorganization & Fixed broken links
> Deleted duplicate textures and materials
> Organize the materials and textures into folders
> Fixed broken links from previous "push"

Note: If ever someone needs to Rename files or Move files around inside UE4, be sure to RightClick on the folder that were affected and Select "Fix Up Redirectors". This will redirect all file references appropriately after the file changes and delete the "hidden" duplicates that UE was still using as reference.